@@ -5,55 +5,43 @@
         "ng": "ng",
         "start": "ng serve",
         "build": "ng build",
+        "watch": "ng build --watch --configuration development",
         "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
+        "lint": "ng lint"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "~12.2.16",
-        "@angular/common": "~12.2.16",
-        "@angular/compiler": "~12.2.16",
-        "@angular/core": "~12.2.16",
-        "@angular/forms": "~12.2.16",
-        "@angular/platform-browser": "~12.2.16",
-        "@angular/platform-browser-dynamic": "~12.2.16",
-        "@angular/router": "~12.2.16",
-        "@kbru/control-props": "^12.0.1",
-        "@ngrx/effects": "^12.5.1",
-        "@ngrx/store": "^12.5.1",
-        "@ngrx/store-devtools": "^12.5.1",
-        "rxjs": "^6.6.3",
-        "tslib": "^2.0.0",
-        "zone.js": "~0.11.4"
+        "@angular/animations": "^18.2.0",
+        "@angular/common": "^18.2.0",
+        "@angular/compiler": "^18.2.0",
+        "@angular/core": "^18.2.0",
+        "@angular/forms": "^18.2.0",
+        "@angular/platform-browser": "^18.2.0",
+        "@angular/platform-browser-dynamic": "^18.2.0",
+        "@angular/router": "^18.2.0",
+        "@ngrx/effects": "^18.0.2",
+        "@ngrx/store": "^18.0.2",
+        "@ngrx/store-devtools": "^18.0.2",
+        "rxjs": "~7.8.0",
+        "tslib": "^2.3.0",
+        "zone.js": "~0.14.10"
     },
     "devDependencies": {
-        "@angular-builders/jest": "^12.1.2",
-        "@angular-devkit/build-angular": "~12.2.17",
-        "@angular-eslint/builder": "12.7.0",
-        "@angular-eslint/eslint-plugin": "12.7.0",
-        "@angular-eslint/eslint-plugin-template": "12.7.0",
-        "@angular-eslint/schematics": "12.7.0",
-        "@angular-eslint/template-parser": "12.7.0",
-        "@angular/cli": "~12.2.17",
-        "@angular/compiler-cli": "~12.2.16",
-        "@types/jest": "^27.5.2",
-        "@types/node": "^12.11.1",
-        "@typescript-eslint/eslint-plugin": "4.28.2",
-        "@typescript-eslint/parser": "4.28.2",
-        "eslint": "^7.26.0",
-        "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-prettier": "^4.0.0",
-        "jasmine-core": "~3.6.0",
-        "jasmine-marbles": "^0.3.1",
-        "jasmine-spec-reporter": "~5.0.0",
-        "jest": "^27.5.1",
-        "jest-marbles": "^2.5.1",
-        "jest-preset-angular": "^11.1.2",
-        "ng-packagr": "^12.2.7",
-        "prettier": "^2.7.1",
-        "protractor": "~7.0.0",
-        "ts-node": "~8.3.0",
-        "typescript": "~4.3.5"
+        "@angular-devkit/build-angular": "^18.2.2",
+        "@angular/cli": "^18.2.2",
+        "@angular/compiler-cli": "^18.2.0",
+        "@types/jasmine": "~5.1.0",
+        "angular-eslint": "18.3.0",
+        "eslint": "^9.9.0",
+        "jasmine-core": "~5.2.0",
+        "karma": "~6.4.0",
+        "karma-chrome-launcher": "~3.2.0",
+        "karma-coverage": "~2.2.0",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.1.0",
+        "ng-packagr": "^18.2.0",
+        "prettier": "^3.3.3",
+        "typescript": "~5.5.2",
+        "typescript-eslint": "8.1.0"
     }
 }
