@@ -1,10 +1,18 @@
 import { Component } from '@angular/core';
+import { ShipmentComponent } from './presentation/shipment/shipment.component';
+import { CitiesModule } from './core/state/cities/cities.module';
+import { ShipmentDataModule } from './core/state/shipment-data/shipment-data.module';
+import { UserDataModule } from './core/state/user-data/user-data.module';
 
 @Component({
     selector: 'demo-root',
-    templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss'],
+    template: '<demo-shipment></demo-shipment>',
+    standalone: true,
+    imports: [
+        ShipmentComponent,
+        CitiesModule,
+        ShipmentDataModule,
+        UserDataModule,
+    ],
 })
-export class AppComponent {
-    title = 'showcases-form';
-}
+export class AppComponent {}
