@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-import { FormControlWithProps, FormGroupWithProps } from '@kbru/control-props';
 import { createEffectAwareForm } from '@kbru/form-effects';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
@@ -8,38 +7,47 @@ import { onZipCodeChangedFormEffect } from '../../form-effects/shipment-form/on-
 import { loadCitiesRequested } from '../../state/cities/cities.actions';
 import { citiesSelector } from '../../state/cities/cities.selectors';
 import { userDataSelector } from '../../state/user-data/user-data.selectors';
-import { ControlProps } from '../../types/control-props.type';
-import { FormGroup } from '../../types/form-group.type';
+import { FormGroup, FormControl, AbstractControl } from '@angular/forms';
+
+class VisibilityAwareFormGroup<
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    T extends { [K in keyof T]: AbstractControl<any, any> },
+> extends FormGroup<T> {
+    public visible = true;
+}
+
+class VisibilityAwareFormControl<T> extends FormControl<T | null> {
+    public visible = true;
+}
+
+class OptionsAwareFormControl<T> extends VisibilityAwareFormControl<T> {
+    public options: T[] = [];
+}
+
+export class ShipmentFormGroup extends VisibilityAwareFormGroup<{
+    address: VisibilityAwareFormGroup<{
+        fullName: VisibilityAwareFormControl<string>;
+        zipCode: VisibilityAwareFormControl<string>;
+        city: OptionsAwareFormControl<string>;
+        street: VisibilityAwareFormControl<string>;
+    }>;
+    method: OptionsAwareFormControl<string>;
+}> {}
 
 @Injectable({ providedIn: 'root' })
 export class ShipmentFormBuilder {
     constructor(private store$: Store) {}
 
-    public build(): Observable<FormGroup> {
-        const form = new FormGroupWithProps<ControlProps>(
-            { visible: true },
-            {
-                address: new FormGroupWithProps(
-                    { visible: true },
-                    {
-                        fullName: new FormControlWithProps(
-                            { visible: true },
-                            ''
-                        ),
-                        zipCode: new FormControlWithProps(
-                            { visible: true },
-                            ''
-                        ),
-                        city: new FormControlWithProps({ visible: true }, ''),
-                        street: new FormControlWithProps({ visible: true }, ''),
-                    }
-                ),
-                method: new FormControlWithProps(
-                    { visible: true, options: ['Standard'] },
-                    'Standard'
-                ),
-            }
-        );
+    public build(): Observable<ShipmentFormGroup> {
+        const form = new ShipmentFormGroup({
+            address: new VisibilityAwareFormGroup({
+                fullName: new VisibilityAwareFormControl(''),
+                zipCode: new VisibilityAwareFormControl(''),
+                city: new OptionsAwareFormControl(''),
+                street: new VisibilityAwareFormControl(''),
+            }),
+            method: new OptionsAwareFormControl<string>('Standard'),
+        });
 
         return createEffectAwareForm(form, [
             onUserDataChangedFormEffect(this.store$.select(userDataSelector)),
