@@ -1,4 +1,3 @@
-import { FormEffect } from '@kbru/form-effects';
 import { Observable, of } from 'rxjs';
 import {
     distinctUntilChanged,
@@ -12,7 +11,8 @@ import {
     CITIES_LOADING,
     CITIES_NOT_LOADED,
 } from '../../state/cities/cities.model';
-import { FormGroup } from '../../types/form-group.type';
+import { FormEffect } from '@kbru/form-effects';
+import { ShipmentFormGroup } from '../../form-builders/shipment/shipment.form-builder';
 
 export const onZipCodeChangedFormEffect =
     (
@@ -20,61 +20,53 @@ export const onZipCodeChangedFormEffect =
             zipCode: string
         ) => Observable<string[] | CitiesLoading | CitiesNotLoaded>,
         loadCities: (zipCode: string) => void
-    ): FormEffect<FormGroup> =>
+    ): FormEffect<ShipmentFormGroup> =>
     formGroup => {
-        const addressGroup = formGroup.get('address');
-        if (!addressGroup) {
-            throw new Error('address not in form');
-        }
-        const zipControl = formGroup.get('address.zipCode');
-        if (!zipControl) {
-            throw new Error('zipCode not in form');
-        }
-        const cityControl = formGroup.get('address.city');
-        if (!cityControl) {
-            throw new Error('city not in form');
-        }
-        const streetControl = formGroup.get('address.street');
-        if (!streetControl) {
-            throw new Error('street not in form');
-        }
+        const addressGroup = formGroup.controls.address;
+        const zipControl = formGroup.controls.address.controls.zipCode;
+        const cityControl = formGroup.controls.address.controls.city;
+        const streetControl = formGroup.controls.address.controls.street;
         return zipControl.valueChanges.pipe(
             startWith(zipControl.value),
             distinctUntilChanged(),
             switchMap(zipCode =>
                 (!zipCode || zipCode.length !== 5
-                    ? of(<string[]>[])
+                    ? of([])
                     : selectCities(zipCode)
                 ).pipe(
-                    map(cities => {
-                        if (CITIES_LOADING === cities) {
-                            return;
+                    map(
+                        (
+                            cities: string[] | CitiesLoading | CitiesNotLoaded
+                        ) => {
+                            if (CITIES_LOADING === cities) {
+                                return;
+                            }
+                            if (CITIES_NOT_LOADED === cities) {
+                                cityControl.disable();
+                                cityControl.options = [];
+                                streetControl.disable();
+                                loadCities(zipCode ?? '');
+                                return;
+                            }
+                            cityControl.visible = true;
+                            cityControl.options = cities;
+                            if (cities.length === 0) {
+                                cityControl.setValue('');
+                                cityControl.disable();
+                                streetControl.disable();
+                                return;
+                            }
+                            if (cities.length === 1) {
+                                cityControl.setValue(cities[0]);
+                                cityControl.visible = false;
+                            }
+                            if (!addressGroup.enabled) {
+                                return;
+                            }
+                            cityControl.enable();
+                            streetControl.enable();
                         }
-                        if (CITIES_NOT_LOADED === cities) {
-                            cityControl.disable();
-                            cityControl.setProp('options', []);
-                            streetControl.disable();
-                            loadCities(zipCode);
-                            return;
-                        }
-                        cityControl.setProp('visible', true);
-                        cityControl.setProp('options', cities);
-                        if (cities.length === 0) {
-                            cityControl.setValue('');
-                            cityControl.disable();
-                            streetControl.disable();
-                            return;
-                        }
-                        if (cities.length === 1) {
-                            cityControl.setValue(cities[0]);
-                            cityControl.setProp('visible', false);
-                        }
-                        if (!addressGroup.enabled) {
-                            return;
-                        }
-                        cityControl.enable();
-                        streetControl.enable();
-                    })
+                    )
                 )
             )
         );
