@@ -1,14 +1,17 @@
 import { NgModule } from '@angular/core';
 import { StoreModule } from '@ngrx/store';
-import { Cities } from './cities.model';
 import { citiesReducer } from './cities.reducer';
 import { EffectsModule } from '@ngrx/effects';
-import { CitiesEffects } from './cities.effects';
+import { loadCitiesEffect } from './cities.effects';
 
 @NgModule({
     imports: [
-        StoreModule.forFeature<Cities>('cities', citiesReducer),
-        EffectsModule.forFeature([CitiesEffects]),
+        StoreModule.forFeature('cities', citiesReducer),
+        EffectsModule.forFeature({ loadCitiesEffect }),
     ],
 })
-export class CitiesModule {}
+export class CitiesModule {
+    constructor() {
+        console.log('CitiesModule.constructor');
+    }
+}
