@@ -1,13 +1,9 @@
-import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { appConfig } from './app/app.config';
+import { AppComponent } from './app/app.component'; // bootstrapApplication(AppModule, appConfig).catch(err => console.error(err));
 
-import { AppModule } from './app/app.module';
-import { environment } from './environments/environment';
+bootstrapApplication(AppComponent, appConfig).catch(err => console.error(err));
 
-if (environment.production) {
-    enableProdMode();
-}
-
-platformBrowserDynamic()
-    .bootstrapModule(AppModule)
-    .catch(err => console.error(err));
+// platformBrowserDynamic()
+//     .bootstrapModule(AppModule)
+//     .catch(err => console.error(err));
